requirements.txt
neo4j
pandas
python-dotenv
langchain[openai,ollama]
openai
numpy
scikit-learn
gensim
rank-bm25
tqdm